pvm general functions and OnTransfer (#118)
* refactor

* on transfer invocation

* refactor + lookup func

* general functions

* refactor

* refactor

* refactor

* refactor

+++ Blockchain/Sources/Blockchain/VMInvocations/HostCall/HostCallFunction.swift
@@ -1,6 +1,8 @@
+import PolkaVM
+
 public protocol HostCallFunction {
     static var identifier: UInt8 { get }
-    static var gasCost: UInt8 { get }
+    static var gasCost: UInt64 { get }
 
     associatedtype Input
     associatedtype Output
